Fixes for the Tour
