+++ app/index.js
@@ -1,4 +1,2 @@
-// "use strict"; require("make-promises-safe"); const app = require("./server.js"); // eslint-disable-line
-
 "use strict"; require("@babel/register"); require("@babel/polyfill");
 module.exports = exports = require("./client.js");
