+++ app/sass/bundle.scss
@@ -21,6 +21,7 @@
   "partials/navigation",
   "partials/mission-statement",
   "partials/modal",
+  "partials/pre",
 
   "layout",
