+++ app/sass/pages/_api.scss
@@ -147,6 +147,10 @@
     margin-bottom: 1rem;
   }
 
+  pre {
+    font-size: 0.8rem;
+  }
+
   table {
     border: 1px solid rgba($white, 0.1);
     border-radius: 0.3rem;
@@ -170,16 +174,6 @@
 
   tr:nth-child(even) {
     background-color: rgba($white, 0.1);
-  }
-
-  pre {
-    margin-bottom: 2rem; padding: 1rem;
-
-    border-radius: 0.3rem;
-    font-size: 0.8rem;
-    line-height: 1.33;
-    overflow-x: auto;
-    overflow-y: hidden;
   }
 }
