+++ app/sass/partials/_github-feed.scss
@@ -50,7 +50,6 @@
     letter-spacing: 0.1rem;
     line-height: 1;
     text-transform: uppercase;
-    width: 100%;
 
     @media (min-width: 1301px) {
       top: 2.15rem; left: 0;
@@ -58,6 +57,7 @@
       color: rgba($black, 0.045);
       font-size: 4rem;
       position: absolute;
+      width: calc(100% - (1rem + 5%));
     }
 
     @media (max-width: 1300px) {
