+++ app/sass/partials/_pre.scss
@@ -1,0 +1,30 @@
+pre {
+  margin-bottom: 2rem; padding: 2rem;
+
+  line-height: 1.33;
+  overflow-x: auto;
+  overflow-y: hidden;
+
+  &:not([class]),
+  &.language-text {
+    background-color: #27283e;
+    color: $white;
+  }
+
+  &.language-yaml {
+    background-color: #27273f;
+    color: #ffe066;
+
+    .atrule {
+      color: #f083ac;
+    }
+
+    .important {
+      color: #ffa94d;
+    }
+
+    .punctuation {
+      color: $white;
+    }
+  }
+}
