+++ documents/playground.md
@@ -1,0 +1,9 @@
+---
+title: Playground
+---
+
+Check out any of the interactive examples to get a feel for the LBRY protocol!
+
+LBRY (pronounced "library") is an application layer protocol, similar to HTTP. However, while HTTP links can direct you to decentralized content, the LBRY protocol *itself* is decentralized.
+
+<Playground/>
